Rename index file for DropDownComponent
Airbitz convention is to name each component's index file as
'index${ComponentName}`
